+++ include/orita.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #ifndef _FILE_ORITA
-#define _FILE_ORITA _FILE_ORITA
+#define _FILE_ORITA
 #include"init.hpp"
 #include"env.h"
 #include"log.hpp"
@@ -8,8 +8,14 @@
 #include"settings.hpp"
 #include"env.hpp"
 #include"process.hpp"
+#include"thread.hpp"
 #include"compile.hpp"
 #include"print.hpp"
 #include"judge.hpp"
 #include"app.hpp"
+void save_config()
+{
+    Files::save();
+    Settings::save();
+}
 #endif