+++ source/check.cpp
@@ -1,6 +1,6 @@
 #pragma once
 #ifndef _COMMAND_CHECK
-#define _COMMAND_CHECK _COMMAND_CHECK
+#define _COMMAND_CHECK
 #include"orita.hpp"
 class Command_check: public App
 {
@@ -71,6 +71,8 @@ class Command_check: public App
             return EXIT_OK;
         }
         unsigned total_sum=std::stoi(get_option("num"));
+        // save config
+        save_config();
         // find file
         if(in==fil()||!in.exists()) {print_result(_in_name,res::type::NF);return EXIT_NOINPUT;}
         if(out==fil()||!out.exists()) {print_result(_out_name,res::type::NF);return EXIT_NOINPUT;}
@@ -104,10 +106,10 @@ class Command_check: public App
             (default_data_dir/"tmp_data").createDirectory();
             INFO("make data dir",add_squo(default_data_dir.path()));
         }
-        catch(Poco::Exception &error)
+        catch(exception &error)
         {
             ERROR("make data dir - fail",add_squo(default_data_dir.path()),add_squo(error.displayText()));
-            throw Poco::Exception("fail make data dir",error.displayText());
+            throw exception("fail make data dir",error.displayText());
         }
         delete print;
         // check